Refactor ddl_extensions into separate modules
To ease understanding and maintenance.
